feat: move formatters to separate folder
